pyproject.toml
[project]
name = "b2speak_api"
version = "0.1.0"
description = ""
authors = [
    {name = "Victor GR"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "pymongo",
    "motor (>=3.7.1,<4.0.0)",
    "azure-storage-blob (>=12.28.0,<13.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
